+++ dvt/tests/test_init.py
@@ -9,8 +9,8 @@ from typer.testing import CliRunner
 
 from devtemplate import __version__
 from devtemplate.cli import app as real_app
-from devtemplate.cli_support import describe_app
 from devtemplate.commands.init import DEFAULT_IMAGE, init
+from devtemplate.describe import describe_app
 
 app = typer.Typer()
 app.command("init")(init)